Update and rename checkmarx-one.js
